refactor component for easier testing
